fix WinHttp tests
