+++ src/libraries/System.Net.Http.WinHttpHandler/tests/FunctionalTests/HttpClientHandlerTestBase.WinHttpHandler.cs
@@ -26,10 +26,27 @@ namespace System.Net.Http.Functional.Tests
             return handler;
         }
 
+        protected WinHttpClientHandler CreateHttpClientHandler() => CreateHttpClientHandler(UseVersion);
+
+        protected static WinHttpClientHandler CreateHttpClientHandler(string useVersionString) =>
+            CreateHttpClientHandler(Version.Parse(useVersionString));
+
         protected static HttpRequestMessage CreateRequest(HttpMethod method, Uri uri, Version version, bool exactVersion = false) =>
             new HttpRequestMessage(method, uri)
             {
                 Version = version
             };
+
+        protected LoopbackServerFactory LoopbackServerFactory => GetFactoryForVersion(UseVersion);
+
+        protected static LoopbackServerFactory GetFactoryForVersion(Version useVersion)
+        {
+            return useVersion.Major switch
+            {
+                2 => Http2LoopbackServerFactory.Singleton,
+                _ => Http11LoopbackServerFactory.Singleton
+            };
+        }
+
     }
 }
