+++ src/libraries/System.Net.Http/tests/FunctionalTests/HttpClientHandlerTestBase.SocketsHttpHandler.cs
@@ -4,6 +4,7 @@
 using System.IO;
 using System.Net.Quic;
 using System.Net.Quic.Implementations;
+using System.Net.Test.Common;
 using System.Reflection;
 using System.Threading.Tasks;
 
@@ -12,6 +13,10 @@ namespace System.Net.Http.Functional.Tests
     public abstract partial class HttpClientHandlerTestBase : FileCleanupTestBase
     {
         protected static bool IsWinHttpHandler => false;
+
+        protected virtual QuicImplementationProvider UseQuicImplementationProvider => null;
+
+        public static bool IsMsQuicSupported => QuicImplementationProviders.MsQuic.IsSupported;
 
         protected static HttpClientHandler CreateHttpClientHandler(Version useVersion = null, QuicImplementationProvider quicImplementationProvider = null)
         {
@@ -33,6 +38,12 @@ namespace System.Net.Http.Functional.Tests
             return handler;
         }
 
+        protected HttpClientHandler CreateHttpClientHandler() => CreateHttpClientHandler(UseVersion, UseQuicImplementationProvider);
+
+        protected static HttpClientHandler CreateHttpClientHandler(string useVersionString) =>
+            CreateHttpClientHandler(Version.Parse(useVersionString));
+
+
         protected static object GetUnderlyingSocketsHttpHandler(HttpClientHandler handler)
         {
             FieldInfo field = typeof(HttpClientHandler).GetField("_underlyingHandler", BindingFlags.Instance | BindingFlags.NonPublic);
@@ -45,6 +56,23 @@ namespace System.Net.Http.Functional.Tests
                 Version = version,
                 VersionPolicy = exactVersion ? HttpVersionPolicy.RequestVersionExact : HttpVersionPolicy.RequestVersionOrLower
             };
+
+        protected LoopbackServerFactory LoopbackServerFactory => GetFactoryForVersion(UseVersion, UseQuicImplementationProvider);
+
+        protected static LoopbackServerFactory GetFactoryForVersion(Version useVersion, QuicImplementationProvider quicImplementationProvider = null)
+        {
+            return useVersion.Major switch
+            {
+#if NETCOREAPP
+#if HTTP3
+                3 => new Http3LoopbackServerFactory(quicImplementationProvider),
+#endif
+                2 => Http2LoopbackServerFactory.Singleton,
+#endif
+                _ => Http11LoopbackServerFactory.Singleton
+            };
+        }
+
     }
 
     internal class VersionHttpClientHandler : HttpClientHandler
